Component for bottom border and move background components to separate directory
