+++ src/client/visualizations/line-chart/base-chart/background/background.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-@import '../../../imports';
+@import '../../../../imports';
 
 .vis-bottom {
   pointer-events: none;
